Moves the pycco_resources package into a module
There's only one file and it can easily be located
within the pycco package.
