+++ pycco/resources.py
